Refactored startup flow.
