+++ Classes/ClientEvents/InteractionCreate.ts
@@ -3,7 +3,7 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
 import { Events, type Interaction } from "discord.js";
-import ClientEvent from "./ClientEvent.ts";
+import ClientEvent from "../ClientEvent.ts";
 
 export default new ClientEvent({
     name: Events.InteractionCreate,
