+++ Classes/ClientEvents/MessageUpdate.ts
@@ -3,7 +3,7 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
 import { ChannelType, Events, type PartialMessage } from "discord.js";
-import ClientEvent from "./ClientEvent.ts";
+import ClientEvent from "../ClientEvent.ts";
 import { editSpectatorMessage } from "../../Modules/messageHandler.ts";
 
 export default new ClientEvent({
